requirements.txt
Flask==0.12.2
Flask-Cors==2.1.2
Flask-Inputs==0.2.0
Flask-Limiter==0.9.3
Flask-RESTful==0.3.5
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.12
WTForms==2.1
Werkzeug==0.11.9
aniso8601==1.1.0
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
dicttoxml==1.6.6
flask-jsontools==0.1.1-0
itsdangerous==0.24
jsonschema==2.5.1
limits==1.1.1
python-dateutil==2.5.3
pytz==2016.4
singledispatch==3.4.0.3
six==1.10.0
# tornado==4.3
# wsgiref==0.1.2
# htmllaundry==2.1
# BeautifulSoup
# pymssql==2.1.2
Flask-Cache
redis
pycrypto
